pyproject.toml
[project]
name = "asepythoncode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.14.17",
    "matplotlib>=3.10.7",
    "ncca-ngl>=0.1.6",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pyopengl>=3.1.10",
    "pyside6>=6.9.2",
    "vulkan>=1.3.275.1",
    "wgpu>=0.25.0",
]

[tool.uv.workspace]
members=["Lecture2","Lecture3","Lecture4","Lab3/Vec3Class/","Lab4/Image/*","Lab4/ImageTests/*","Lab5/*","GraphicsDemos/*","Lab6/*"]
exclude =["*"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ipykernel>=6.16.2",

]


[tool.marimo.display]
default_width = "full"

[tool.ruff]
exclude = [
    "*.ipynb",   # ignore all notebooks
]
